Download_unpack() should use **kwargs
download_unpack() ends up calling a function at the end of it.
There are arguments of which are only needed for such function call.
If we use kwargs we won't have to change the signature of
download_unpack() every time we add some new feature to the function
we end up calling.